requirements.txt
pandas==2.2.2
numpy==2.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.1
six==1.16.0
openpyxl==3.1.5
et-xmlfile==1.1.0
requests==2.32.3
charset-normalizer==3.3.2
idna==3.10
urllib3==2.2.3
certifi==2024.8.30
python-dotenv==1.0.1